feat: Replace game code with boilerplate
BREAKING CHANGE: Remove old game code, replace with boilerplate
